+++ .eslintrc.js
@@ -1,24 +1,11 @@
-/**
- * @remarks
- * This could just use `extends`
- * but making a monorepo module a dependency in the root
- * causes an extra copy to be saved to node-modules that isn't
- * symlinked, which we don't want. This should eventually just
- * move to a shared private module instead of having this in the
- * project root
- */
-const defaultConfig = require('./packages/eslint-config');
-
 module.exports = {
-  ...defaultConfig,
   root: true,
 
+  extends: [require.resolve('./packages/eslint-config')],
+
+  plugins: ['local-rules'],
+
   overrides: [
-    ...defaultConfig.overrides,
-    {
-      files: ['*.mdx'],
-      parser: 'eslint-mdx',
-    },
     {
       files: ['**/typings/*', '*.d.ts'],
       parserOptions: {
@@ -45,6 +32,7 @@ module.exports = {
               'ImportDeclaration[source.value=/^((?!module.scss).)*(.scss)$/]',
           },
         ],
+        'local-rules/gamut-import-paths': 'error',
       },
     },
   ],