+++ packages/gamut-labs/src/experimental/AppHeader/AppHeaderElements/AppHeaderLogo/index.tsx
@@ -1,4 +1,3 @@
-import { Anchor } from '@codecademy/gamut';
 import styled from '@emotion/styled';
 import React from 'react';
 
@@ -6,7 +5,7 @@ import { Logo } from '../../../../brand/Logo';
 import { focusStyles } from '../SharedStyles';
 import { AppHeaderClickHandler, AppHeaderLogoItem } from '../types';
 
-const LogoButton = styled(Anchor)`
+const LogoButton = styled.a`
   display: inline-block;
   padding: 0.375rem 0;
   ${focusStyles}
@@ -32,7 +31,7 @@ export const AppHeaderLogo: React.FC<AppHeaderLogoProps> = ({
       onClick={(event) => action(event, item)}
       href={item.href}
     >
-      {<StyledLogo type={item.pro ? 'proMono' : 'default'} height={27} />}
+      <StyledLogo type={item.pro ? 'proMono' : 'default'} height={27} />
     </LogoButton>
   );
 };