+++ packages/gamut-labs/src/experimental/GlobalHeader/GlobalHeaderItems.tsx
@@ -1,6 +1,7 @@
 import {
   AccountingCoinsIcon,
   BookFlipPageIcon,
+  BriefcaseIcon,
   CommunityIcon,
   GearIcon,
   HouseEntranceIcon,
@@ -12,12 +13,12 @@ import {
 import { ReactNode } from 'react';
 
 import {
-  AppHeaderDropdownItem,
   AppHeaderFillButtonItem,
   AppHeaderLinkItem,
   AppHeaderLogoItem,
   AppHeaderProfileDropdownItem,
   AppHeaderRenderElementItem,
+  AppHeaderSimpleDropdownItem,
   AppHeaderTextButtonItem,
 } from '../AppHeader/AppHeaderElements/types';
 import { User } from './types';
@@ -42,6 +43,7 @@ export const proLogo: AppHeaderLogoItem = {
 
 export const myHome: AppHeaderLinkItem = {
   dataTestId: 'header-home',
+  dataIntellimizeId: 'header-home',
   icon: HouseEntranceIcon,
   id: 'my-home',
   text: 'My Home',
@@ -52,15 +54,16 @@ export const myHome: AppHeaderLinkItem = {
 
 export const courseCatalog: AppHeaderLinkItem = {
   dataTestId: 'header-catalog',
+  dataIntellimizeId: 'header-catalog',
   icon: BookFlipPageIcon,
   id: 'course-catalog',
-  text: 'Course Catalog',
+  text: 'Catalog',
   href: '/catalog',
   trackingTarget: 'topnav_catalog',
   type: 'link',
 };
 
-export const resourcesDropdown: AppHeaderDropdownItem = {
+export const resourcesDropdown: AppHeaderSimpleDropdownItem = {
   icon: NotebookIcon,
   id: 'resources',
   text: 'Resources',
@@ -73,6 +76,13 @@ export const resourcesDropdown: AppHeaderDropdownItem = {
       type: 'link',
     },
     {
+      id: 'projects',
+      href: '/projects',
+      trackingTarget: 'topnav_resources_projects',
+      text: 'Projects',
+      type: 'link',
+    },
+    {
       id: 'articles',
       href: '/articles',
       trackingTarget: 'topnav_resources_articles',
@@ -82,6 +92,7 @@ export const resourcesDropdown: AppHeaderDropdownItem = {
     {
       id: 'blog',
       href: 'https://news.codecademy.com/',
+      newTab: true,
       trackingTarget: 'topnav_resources_blog',
       text: 'Blog',
       type: 'link',
@@ -91,7 +102,7 @@ export const resourcesDropdown: AppHeaderDropdownItem = {
   type: 'dropdown',
 };
 
-export const communityDropdown: AppHeaderDropdownItem = {
+export const communityDropdown: AppHeaderSimpleDropdownItem = {
   icon: CommunityIcon,
   id: 'community',
   text: 'Community',
@@ -100,6 +111,7 @@ export const communityDropdown: AppHeaderDropdownItem = {
       id: 'forums',
       href: 'https://discuss.codecademy.com/',
       trackingTarget: 'topnav_community_forums',
+      newTab: true,
       text: 'Forums',
       type: 'link',
     },
@@ -114,6 +126,7 @@ export const communityDropdown: AppHeaderDropdownItem = {
     {
       id: 'chapters',
       href: 'https://community.codecademy.com/',
+      newTab: true,
       trackingTarget: 'topnav_community_chapters',
       text: 'Chapters',
       type: 'link',
@@ -130,7 +143,7 @@ export const communityDropdown: AppHeaderDropdownItem = {
   type: 'dropdown',
 };
 
-export const pricingDropdown: AppHeaderDropdownItem = {
+export const pricingDropdown: AppHeaderSimpleDropdownItem = {
   icon: AccountingCoinsIcon,
   id: 'pricing',
   text: 'Pro Pricing',
@@ -155,7 +168,9 @@ export const pricingDropdown: AppHeaderDropdownItem = {
 };
 
 export const forBusiness: AppHeaderLinkItem = {
+  icon: BriefcaseIcon,
   id: 'for-business',
+  dataIntellimizeId: 'header-business',
   trackingTarget: 'topnav_business',
   text: 'For Business',
   href: '/business',
@@ -209,7 +224,7 @@ const profileMyHome: AppHeaderLinkItem = {
   type: 'link',
 };
 
-const profileBusiness: AppHeaderLinkItem = {
+const profileBusinessAccount: AppHeaderLinkItem = {
   id: 'business',
   icon: PieLineGraphIcon,
   href: '/business/plans',
@@ -230,6 +245,7 @@ const profileHelpCenter: AppHeaderLinkItem = {
 
 const profileAdmin: AppHeaderLinkItem = {
   id: 'admin',
+  dataTestId: 'admin-link',
   href: '/admin',
   trackingTarget: 'avatar_admin',
   text: 'Admin',
@@ -247,6 +263,7 @@ const profileCustomerSupport: AppHeaderLinkItem = {
 const profileReportBug: AppHeaderLinkItem = {
   id: 'report-bug',
   href: 'https://codecademy.atlassian.net/servicedesk/customer/portal/9',
+  newTab: true,
   trackingTarget: 'avatar_report_bug',
   text: 'Report a Bug [ADMIN]',
   type: 'link',
@@ -260,41 +277,23 @@ const profileLogOut: AppHeaderLinkItem = {
   type: 'link',
 };
 
-export const freeProfile = (user: User): AppHeaderProfileDropdownItem => {
+export const freeProfile = (
+  user: User,
+  isMobile?: boolean
+): AppHeaderProfileDropdownItem => {
+  const topSection = [profileMyProfile, profileAccount, profileMyHome];
+  if (!isMobile && user.isAccountManager) {
+    topSection.push(profileBusinessAccount);
+  }
+  topSection.push(profileHelpCenter);
+
+  const bottomSection = [profileLogOut];
+
+  const popover = [topSection, bottomSection];
+
   return {
     avatar: user.avatar,
-    displayName: user.displayName,
-    id: 'profile',
-    text: 'Profile',
-    popover: [
-      [profileMyProfile, profileAccount, profileMyHome, profileHelpCenter],
-      [profileLogOut],
-    ],
-    trackingTarget: 'topnav_profile',
-    type: 'profile-dropdown',
-  };
-};
-
-export const proProfile = (user: User): AppHeaderProfileDropdownItem => {
-  const popover = [];
-  popover.push([
-    profileMyProfile,
-    profileAccount,
-    profileMyHome,
-    profileBusiness,
-    profileHelpCenter,
-  ]);
-
-  const adminSection = [];
-  user.isAdmin && adminSection.push(profileAdmin);
-  user.isCustomerSupport && adminSection.push(profileCustomerSupport);
-  user.isAdmin && adminSection.push(profileReportBug);
-  popover.push(adminSection);
-
-  popover.push([profileLogOut]);
-  return {
-    avatar: user.avatar,
-    displayName: user.displayName,
+    userDisplayName: user.displayName,
     id: 'profile',
     text: 'Profile',
     popover,
@@ -303,11 +302,68 @@ export const proProfile = (user: User): AppHeaderProfileDropdownItem => {
   };
 };
 
-export const upgradeToPro: AppHeaderFillButtonItem = {
-  id: 'upgrade-to-pro',
+export const proProfile = (
+  user: User,
+  isMobile?: boolean
+): AppHeaderProfileDropdownItem => {
+  const topSection = [profileMyProfile, profileAccount, profileMyHome];
+  if (!isMobile && (user.isAccountManager || user.isAdmin)) {
+    topSection.push(profileBusinessAccount);
+  }
+  topSection.push(profileHelpCenter);
+
+  const middleSection = [];
+  if (user.isAdmin) {
+    middleSection.push(profileAdmin);
+  }
+  if (user.isCustomerSupport) {
+    middleSection.push(profileCustomerSupport);
+  }
+  if (user.isAdmin) {
+    middleSection.push(profileReportBug);
+  }
+
+  const bottomSection = [profileLogOut];
+
+  const popover = [topSection, middleSection, bottomSection];
+
+  return {
+    avatar: user.avatar,
+    userDisplayName: user.displayName,
+    id: 'profile',
+    text: 'Profile',
+    popover,
+    trackingTarget: 'topnav_profile',
+    type: 'profile-dropdown',
+  };
+};
+
+export const tryProForFree = (
+  checkoutUrl?: string
+): AppHeaderFillButtonItem => ({
+  dataTestId: 'upgrade-link',
+  id: 'try-pro',
   text: 'Try Pro For Free',
-  href: '/pro/membership',
+  href: checkoutUrl || '/pro/membership',
   trackingTarget: 'topnav_pro_trial',
+  type: 'fill-button',
+});
+
+export const upgradeToPro: AppHeaderFillButtonItem = {
+  dataTestId: 'upgrade-link',
+  id: 'upgrade-to-pro',
+  text: 'Upgrade to Pro',
+  href: '/pro/membership',
+  trackingTarget: 'topnav_pro_upgrade',
+  type: 'fill-button',
+};
+
+export const unpausePro: AppHeaderFillButtonItem = {
+  dataTestId: 'unpause-link',
+  id: 'unpause-pro',
+  text: 'Unpause Now',
+  href: '/account/billing',
+  trackingTarget: 'topnav_pro_unpause',
   type: 'fill-button',
 };
 
@@ -318,6 +374,7 @@ export const login: AppHeaderTextButtonItem = {
   href: '/login',
   trackingTarget: 'topnav_login',
   type: 'text-button',
+  redirect: true,
 };
 
 export const signUp: AppHeaderFillButtonItem = {
@@ -327,4 +384,5 @@ export const signUp: AppHeaderFillButtonItem = {
   href: '/register',
   trackingTarget: 'topnav_signup',
   type: 'fill-button',
+  redirect: true,
 };