+++ packages/gamut-labs/src/experimental/GlobalHeader/__tests__/GlobalHeader-test.tsx
@@ -15,6 +15,8 @@ import {
   pricingDropdown,
   resourcesDropdown,
   signUp,
+  tryProForFree,
+  unpausePro,
   upgradeToPro,
 } from '../GlobalHeaderItems';
 import { User } from '../types';
@@ -39,14 +41,20 @@ const anonLandingHeaderProps: GlobalHeaderProps = {
 
 const anonLoginHeaderProps: GlobalHeaderProps = {
   action,
-  renderSearch: () => <IconButton icon={SearchIcon} />,
+  renderSearch: {
+    desktop: () => <IconButton icon={SearchIcon} />,
+    mobile: () => <IconButton icon={SearchIcon} />,
+  },
   type: 'anon',
   variant: 'login',
 };
 
 const anonSignUpHeaderProps: GlobalHeaderProps = {
   action,
-  renderSearch: () => <IconButton icon={SearchIcon} />,
+  renderSearch: {
+    desktop: () => <IconButton icon={SearchIcon} />,
+    mobile: () => <IconButton icon={SearchIcon} />,
+  },
   type: 'anon',
   variant: 'signup',
 };
@@ -57,16 +65,54 @@ const freeHeaderProps: GlobalHeaderProps = {
   user,
 };
 
+const freeCustomCheckoutUrlHeaderProps: GlobalHeaderProps = {
+  action,
+  type: 'free',
+  user: {
+    proTrialCheckoutUrl: 'test-url',
+    ...user,
+  },
+};
+
+const freeCompletedTrialHeaderProps: GlobalHeaderProps = {
+  action,
+  type: 'free',
+  user: {
+    showProUpgrade: true,
+    ...user,
+  },
+};
+
 const proHeaderProps: GlobalHeaderProps = {
   action,
   type: 'pro',
   user,
 };
 
+const proPausedHeaderProps: GlobalHeaderProps = {
+  action,
+  type: 'pro',
+  user: {
+    isPaused: true,
+    ...user,
+  },
+};
+
+const loadingHeaderProps: GlobalHeaderProps = {
+  action,
+  type: 'loading',
+};
+
 const renderElementProps: GlobalHeaderProps = {
   action,
-  renderSearch: () => <IconButton icon={SearchIcon} />,
-  renderNotifications: () => <IconButton icon={BellIcon} />,
+  renderSearch: {
+    desktop: () => <IconButton icon={SearchIcon} />,
+    mobile: () => <IconButton icon={SearchIcon} />,
+  },
+  renderNotifications: {
+    desktop: () => <IconButton icon={BellIcon} />,
+    mobile: () => <IconButton icon={BellIcon} />,
+  },
   type: 'pro',
   user,
 };
@@ -85,36 +131,40 @@ describe('GlobalHeader', () => {
       renderGlobalHeader(anonHeaderProps);
     });
 
+    /* since we're using css to toggle the display of the header between desktop & mobile, these tests check for visibility of elements
+     & use 'getAllByTestId' b/c there will be duplicate elements in the DOM (since mobile & desktop render some of the same app header items) */
     test('logo', () => {
-      screen.getByTestId('header-logo');
+      const logoElements = screen.getAllByTestId('header-logo');
+      expect(logoElements[0]).not.toBeVisible();
+      expect(logoElements[1]).toBeVisible();
     });
 
     test('courseCatalog', () => {
-      screen.getByText(courseCatalog.text);
+      screen.getAllByText(courseCatalog.text);
     });
 
     test('resourcesDropdown', () => {
-      screen.getByText(resourcesDropdown.text);
+      screen.getAllByText(resourcesDropdown.text);
     });
 
     test('communityDropdown', () => {
-      screen.getByText(communityDropdown.text);
+      screen.getAllByText(communityDropdown.text);
     });
 
     test('plansPricingDropdown', () => {
-      screen.getByText(pricingDropdown.text);
+      screen.getAllByText(pricingDropdown.text);
     });
 
     test('forEnterprise', () => {
-      screen.getByText(forBusiness.text);
+      screen.getAllByText(forBusiness.text);
     });
 
     test('login', () => {
-      screen.getByText(login.text);
+      screen.getAllByText(login.text);
     });
 
     test('signup', () => {
-      screen.getByText(signUp.text);
+      screen.getAllByText(signUp.text);
     });
   });
 
@@ -129,7 +179,7 @@ describe('GlobalHeader', () => {
       });
 
       test('shows login', () => {
-        screen.getByText(login.text);
+        screen.getAllByText(login.text);
       });
 
       test('does not show signup', () => {
@@ -151,7 +201,7 @@ describe('GlobalHeader', () => {
       });
 
       test('shows signup', () => {
-        screen.getByText(signUp.text);
+        screen.getAllByText(signUp.text);
       });
     });
 
@@ -161,11 +211,11 @@ describe('GlobalHeader', () => {
       });
 
       test('shows search', () => {
-        screen.getByTitle('Search Icon');
+        screen.getAllByTitle('Search Icon');
       });
 
       test('shows login', () => {
-        screen.getByText(login.text);
+        screen.getAllByText(login.text);
       });
 
       test('does not show sign up', () => {
@@ -175,74 +225,109 @@ describe('GlobalHeader', () => {
   });
 
   describe('free users', () => {
-    beforeEach(() => {
-      renderGlobalHeader(freeHeaderProps);
+    describe('default', () => {
+      beforeEach(() => {
+        renderGlobalHeader(freeHeaderProps);
+      });
+
+      test('logo', () => {
+        screen.getAllByTestId('header-logo');
+      });
+
+      test('myHome', () => {
+        screen.getByText(myHome.text);
+      });
+
+      test('courseCatalog', () => {
+        screen.getByText(courseCatalog.text);
+      });
+
+      test('resourcesDropdown', () => {
+        screen.getByText(resourcesDropdown.text);
+      });
+
+      test('communityDropdown', () => {
+        screen.getByText(communityDropdown.text);
+      });
+
+      test('plansPricingDropdown', () => {
+        screen.getByText(pricingDropdown.text);
+      });
+
+      test('forEnterprise', () => {
+        screen.getByText(forBusiness.text);
+      });
+
+      test('profileDropdown', () => {
+        screen.getByTestId('avatar');
+      });
+
+      test('tryProForFree', () => {
+        screen.getByText(tryProForFree().text);
+      });
     });
 
-    test('logo', () => {
-      screen.getByTestId('header-logo');
+    describe('custom checkout url', () => {
+      test('tryProForFree', () => {
+        renderGlobalHeader(freeCustomCheckoutUrlHeaderProps);
+        screen.getByText(tryProForFree().text);
+      });
     });
 
-    test('myHome', () => {
-      screen.getByText(myHome.text);
-    });
-
-    test('courseCatalog', () => {
-      screen.getByText(courseCatalog.text);
-    });
-
-    test('resourcesDropdown', () => {
-      screen.getByText(resourcesDropdown.text);
-    });
-
-    test('communityDropdown', () => {
-      screen.getByText(communityDropdown.text);
-    });
-
-    test('plansPricingDropdown', () => {
-      screen.getByText(pricingDropdown.text);
-    });
-
-    test('forEnterprise', () => {
-      screen.getByText(forBusiness.text);
-    });
-
-    test('profileDropdown', () => {
-      screen.getByTestId('avatar');
-    });
-
-    test('upgradeToPro', () => {
-      screen.getByText(upgradeToPro.text);
+    describe('has completed trial', () => {
+      test('upgradeToPro', () => {
+        renderGlobalHeader(freeCompletedTrialHeaderProps);
+        screen.getByText(upgradeToPro.text);
+      });
     });
   });
 
   describe('pro users', () => {
+    describe('default', () => {
+      beforeEach(() => {
+        renderGlobalHeader(proHeaderProps);
+      });
+
+      test('proLogo', () => {
+        screen.getAllByTestId('header-pro-logo');
+      });
+
+      test('myHome', () => {
+        screen.getByText(myHome.text);
+      });
+
+      test('courseCatalog', () => {
+        screen.getByText(courseCatalog.text);
+      });
+
+      test('resourcesDropdown', () => {
+        screen.getByText(resourcesDropdown.text);
+      });
+
+      test('communityDropdown', () => {
+        screen.getByText(communityDropdown.text);
+      });
+
+      test('profileDropdown', () => {
+        screen.getByTestId('avatar');
+      });
+    });
+
+    describe('is paused', () => {
+      test('unpause', () => {
+        renderGlobalHeader(proPausedHeaderProps);
+        screen.getByText(unpausePro.text);
+      });
+    });
+  });
+
+  describe('loading', () => {
     beforeEach(() => {
-      renderGlobalHeader(proHeaderProps);
+      renderGlobalHeader(loadingHeaderProps);
     });
 
-    test('proLogo', () => {
-      screen.getByTestId('header-pro-logo');
-    });
-
-    test('myHome', () => {
-      screen.getByText(myHome.text);
-    });
-
-    test('courseCatalog', () => {
-      screen.getByText(courseCatalog.text);
-    });
-
-    test('resourcesDropdown', () => {
-      screen.getByText(resourcesDropdown.text);
-    });
-
-    test('communityDropdown', () => {
-      screen.getByText(communityDropdown.text);
-    });
-
-    test('profileDropdown', () => {
-      screen.getByTestId('avatar');
+    test('logo', () => {
+      screen.getAllByTestId('header-logo');
     });
   });
 
@@ -256,13 +341,13 @@ describe('GlobalHeader', () => {
     });
 
     test('notifications', () => {
-      screen.getByTitle('Bell Icon');
+      screen.getAllByTitle('Bell Icon');
     });
   });
 
   test('fires action() upon clicking an element', () => {
     renderGlobalHeader(renderElementProps);
-    screen.getAllByRole('button')[0].click();
+    screen.getAllByRole('link')[0].click();
     expect(action).toHaveBeenCalled();
   });
 });