+++ packages/gamut-styles/src/theme/theme.tsx
@@ -1,0 +1,13 @@
+import * as tokens from '../variables';
+
+export const theme = {
+  boxShadows: tokens.boxShadows,
+  breakpoints: tokens.mediaQueries,
+  fontSize: tokens.fontSize,
+  fontFamily: tokens.fontFamily,
+  lineHeight: tokens.lineHeight,
+  fontWeight: tokens.fontWeight,
+  colors: tokens.colors,
+  spacing: tokens.spacing,
+  elements: tokens.elements,
+} as const;