+++ packages/gamut/src/GridForm/index.tsx
@@ -9,6 +9,11 @@ import { GridFormField } from './types';
 
 export * from './types';
 
+const defaultColumnGap = {
+  xs: 'sm',
+  sm: 'lg',
+} as const;
+
 export type GridFormProps<Values extends {}> = {
   children?: React.ReactNode;
   className?: string;
@@ -27,6 +32,11 @@ export type GridFormProps<Values extends {}> = {
    * Function called with field values on submit, if all validations have passed.
    */
   onSubmit: SubmitHandler<Values>;
+
+  /**
+   * Show asterisks next to required fields.
+   */
+  showRequired?: boolean;
 
   /**
    * Layout grid row gap override.
@@ -56,12 +66,13 @@ export function GridForm<
 >({
   children,
   className,
-  columnGap = 'lg',
+  columnGap = defaultColumnGap,
   fields = [],
   onSubmit,
   rowGap = 'md',
   submit,
   validation = 'onSubmit',
+  showRequired = false,
 }: GridFormProps<Values>) {
   const {
     errors,
@@ -94,6 +105,9 @@ export function GridForm<
 
           const isFirstError = !pastFirstError && errorMessage !== undefined;
           pastFirstError = pastFirstError || isFirstError;
+          const requiredBoolean = !!(
+            field.validation && field.validation.required
+          );
 
           return (
             <GridFormInputGroup
@@ -103,6 +117,8 @@ export function GridForm<
               key={field.name}
               register={register}
               setValue={setValue}
+              required={requiredBoolean}
+              showRequired={showRequired}
             />
           );
         })}