+++ packages/tracking/src/types.ts
@@ -51,9 +51,12 @@ export type EventDataTypes = {
     query: BaseEventAnyData;
     result: BaseEventAnyData;
   };
+  // page events for email marketing
+  // potentially to be removed once we clean up user_visit event
   page: {
     career_path_visited: PagePathVisitedData;
     skill_path_visited: PagePathVisitedData;
+    course_page_visited: CoursePageVisitedData;
   };
 };
 
@@ -85,8 +88,6 @@ export type BaseEventAnyData = BaseEventData & {
 export type TrackingOptions = {
   /** tells backend not to merge user-identifying data onto the event payload */
   gdprSafe?: boolean;
-  /** tells frontend to send details to fullstory, which only accepts client-side events */
-  fullstory?: boolean;
 };
 
 /**
@@ -168,6 +169,7 @@ export type UserClickData = UserSharedData & {
   course_slug?: string;
   course_progress?: number;
   assessment_id?: string;
+  container_slugs?: string[];
 };
 
 /**
@@ -226,7 +228,12 @@ export type UseUserResponse = {
   status: string;
 };
 
-export declare type PagePathVisitedData = BaseEventData & {
+export type PagePathVisitedData = BaseEventData & {
   path_id: string;
   path_full_title: string;
+};
+
+export type CoursePageVisitedData = BaseEventData & {
+  course_id: string;
+  course_full_title: string;
 };